Fix QNX, BSD and WIN platform dependant errors

+++ include/boost/dll/detail/elf_info.hpp
@@ -172,7 +172,7 @@ private:
 
         sections.resize(static_cast<std::size_t>(section_names_section.sh_size));
         f_.seekg(section_names_section.sh_offset);
-        read_raw(sections[0], section_names_section.sh_size);
+        read_raw(sections[0], static_cast<std::size_t>(section_names_section.sh_size));
     }
 
     void symbols_text(std::vector<symbol_t>& symbols, std::vector<char>& text) {
@@ -188,14 +188,14 @@ private:
 
                 const boost::filesystem::ifstream::pos_type pos = f_.tellg();
                 f_.seekg(section.sh_offset);
-                read_raw(symbols[0], section.sh_size - (section.sh_size % sizeof(symbol_t)) );
+                read_raw(symbols[0], static_cast<std::size_t>(section.sh_size - (section.sh_size % sizeof(symbol_t))) );
                 f_.seekg(pos);
             } else if (section.sh_type == SHT_STRTAB_) {
                 text.resize(static_cast<std::size_t>(section.sh_size));
 
                 const boost::filesystem::ifstream::pos_type pos = f_.tellg();
                 f_.seekg(section.sh_offset);
-                read_raw(text[0], section.sh_size);
+                read_raw(text[0], static_cast<std::size_t>(section.sh_size));
                 f_.seekg(pos);
             }
         }
